解决自定义替身第一人称的显示问题

+++ src/main/java/com/lh_lshen/mcbbs/huajiage/client/resources/CustomResourceLoader.java
@@ -79,15 +79,15 @@ public class CustomResourceLoader {
                     // 尝试加载模型
                     ModelStandJson modelJson = loadModel(standModelItem.getModel());
                     if (standModelItem.getTransfer()!=null && standModelItem.getTransfer().size()>=3) {
-                        modelJson.setPositions((float) standModelItem.getTransfer().get(0),
-                                               (float) standModelItem.getTransfer().get(1),
-                                               (float) standModelItem.getTransfer().get(2));
+                        modelJson.setPositions(standModelItem.getTransfer().get(0),
+                                               standModelItem.getTransfer().get(1),
+                                               standModelItem.getTransfer().get(2));
                     }
                     if (standModelItem.getRotation()!=null && standModelItem.getRotation().size()>=4) {
-                        modelJson.setRotations((float) standModelItem.getRotation().get(0),
-                                               (float) standModelItem.getRotation().get(1),
-                                               (float) standModelItem.getRotation().get(2),
-                                               (float) standModelItem.getRotation().get(3));
+                        modelJson.setRotations(standModelItem.getRotation().get(0),
+                                               standModelItem.getRotation().get(1),
+                                               standModelItem.getRotation().get(2),
+                                               standModelItem.getRotation().get(3));
                     }
                     // 加载动画
                     @Nullable List<Object> animations = CustomJsAnimationManger.getCustomAnimation(standModelItem);
